feat: getValidMenuForReview API 수정에 따른 변경
