+++ app/src/main/java/com/eatssu/android/domain/repository/MealRepository.kt
@@ -1,7 +1,5 @@
 package com.eatssu.android.domain.repository
 
-import com.eatssu.android.data.dto.response.BaseResponse
-import com.eatssu.android.data.dto.response.MenuOfMealResponse
 import kotlinx.coroutines.flow.Flow
 
 interface MealRepository {
@@ -15,11 +13,4 @@ interface MealRepository {
         time: String,
     ): Flow<List<List<String>>>
 
-
-    /**
-     * MealId를 이용해서 Menu를 찾기 api
-     */
-    suspend fun getMenuInfoByMealId(
-        mealId: Long,
-    ): Flow<BaseResponse<MenuOfMealResponse>>
 }
