+++ app/src/main/java/com/eatssu/android/domain/repository/ReviewRepository.kt
@@ -45,4 +45,11 @@ interface ReviewRepository {
     suspend fun getImageString(
         file: File
     ): String
+
+    /**
+     * MealId를 이용해서 Menu를 찾기 api (+ 리뷰 작성 가능한 메뉴 조회 v2)
+     */
+    suspend fun getMenuInfoByMealId(
+        mealId: Long,
+    ): List<Pair<Long, String>>
 }
